Synaptics: Fixed typo in the driver
It looks like a unplanned deviation from Synaptics, altough the consistency of it (e.g in the same typo both in function definition and call) makes me somehow wonder or even doubt it. Only compiling and testing will reveal whether it indeed is a typo or rather a planned deviation.